requirements.txt
# Core libraries
pandas
numpy
scikit-learn
matplotlib
seaborn

# ETL components
yfinance
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
apache-airflow
ta
ccxt

# ML components
tensorflow
mlflow


# API & Dashboard
fastapi
uvicorn
plotly
dash

# Development tools
pytest
black
isort
flake8
python-dotenv